Weights controller and ciew pages

+++ db/migrate/20170827184203_create_weights.rb
@@ -2,7 +2,7 @@ class CreateWeights < ActiveRecord::Migration[5.1]
   def change
     create_table :weights do |t|
       t.integer :user_id
-      t.integer :amount
+      t.float :amount
       t.datetime :date
       t.timestamps
     end
